requirements.txt
numpy
pillow
onnxruntime
starlette
python-multipart
uvicorn
